+++ mayan/apps/storage/model_mixins.py
@@ -30,6 +30,24 @@ class DatabaseFileModelMixin(models.Model):
             'name': self.file.name
         }
 
+        # Close the self.file object as Django generates a new descriptor
+        # when the file field is accessed.
+        # From django/db/models/fields/files.py.
+        """
+        The descriptor for the file attribute on the model instance. Return a
+        FieldFile when accessed so you can write code like::
+
+            >>> from myapp.models import MyModel
+            >>> instance = MyModel.objects.get(pk=1)
+            >>> instance.file.size
+
+        Assign a file object on assignment so you can do::
+
+            >>> with open('/path/to/hello.world', 'r') as f:
+            ...     instance.file = File(f)
+        """
+        self.file.close()
+
         default_kwargs.update(**kwargs)
 
         # Ensure the caller cannot specify an alternate filename.